Move variant util functions and rename radio price component
